refactor : 코드 리펙터링
