+++ src/main/java/ruby/ioccontainer/annotation/CustomAutowired.java
@@ -5,6 +5,10 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+/**
+ * CustomAutowired
+ * - 해당 애너테이션이 붙은 필드에 빈 객체를 주입
+ */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.FIELD})
 public @interface CustomAutowired {
