+++ src/test/java/ruby/testioc/CustomComponentScanTest.java
@@ -1,4 +1,4 @@
-package ruby.container;
+package ruby.testioc;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
