+++ src/test/java/ruby/testioc/TestApp.java
@@ -1,4 +1,4 @@
-package ruby.container;
+package ruby.testioc;
 
 import ruby.ioccontainer.annotation.CustomComponentScan;
