+++ src/test/java/ruby/testioc/TestBean2.java
@@ -1,4 +1,4 @@
-package ruby.container;
+package ruby.testioc;
 
 import ruby.ioccontainer.annotation.CustomComponent;
